Make compatible with docker.

+++ dev_contain/clean.py
@@ -18,6 +18,7 @@ import argparse
 import sys
 import subprocess
 import shlex
+import dev_contain.common as common
 
 def clean(in_args):
     parser = argparse.ArgumentParser(prog=sys.argv[0]+' clean', description='Prune system of uneeded containers and images.')
@@ -25,20 +26,28 @@ def clean(in_args):
     parser.add_argument('--container', '-c', help='Stop and remove a running container.')
     args = parser.parse_args(in_args)
 
+    builder = common.get_builder()
+    manager = common.get_manager()
+
     if(args.buildah):
-        run_and_log('Removing buildah containers.', 'buildah rm --all')
-        run_and_log('Pruning buildah images.', 'buildah rmi --prune')
+        if builder == 'buildah':
+            run_and_log('Removing buildah containers.', 'buildah rm --all')
+            run_and_log('Pruning buildah images.', 'buildah rmi --prune')
+        else:
+            print('Builder is {}. I only know how to clean up for buildah.'.format(builder))
     if(args.container):
         # Is container running?
-        check_command = 'podman inspect ' + args.container  +' -f "{{ .State.Running }}"'
+        check_command = manager + ' inspect ' + args.container  +' -f "{{ .State.Running }}"'
         check_result = subprocess.Popen(shlex.split(check_command), stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         is_running = check_result.stdout.read().decode('ascii').strip() == 'true'
         # If so stop it.
         if(is_running):
-            stop_command = 'podman stop --timeout 1 {}'.format(args.container)
+            stop_command = '{manager} stop {container}'.format(manager=manager,
+                                                               container=args.container)
             run_and_log('Stopping container.', stop_command)
         # rm container.
-        remove_command = 'podman rm {}'.format(args.container)
+        remove_command = '{manager} rm {container}'.format(manager=manager,
+                                                           container=args.container)
         run_and_log('Stopping container.', remove_command)
 
 def run_and_log(comment, command):
